requirements.txt
Django>=4.2,<6.0
djangorestframework
python-dotenv
psycopg2-binary
gunicorn
tiktoken
openai
qdrant-client
pdfplumber

